# openapi.yaml
openapi: 3.0.0
info:
  title: API de Charadas do Tio João
  description: Uma API para gerenciamento de charadas com integração ao Firebase/Firestore.
  version: 1.0.0

servers:
  - url: http://localhost:5000
    description: Servidor de Desenvolvimento Local

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Insira o token JWT gerado na rota /login.

  schemas:
    Charada:
      type: object
      properties:
        id:
          type: integer
          example: 1
        pergunta:
          type: string
          example: "O que é, o que é?"
        resposta:
          type: string
          example: "É o café."

    LoginRequest:
      type: object
      required:
        - usuario
        - senha
      properties:
        usuario:
          type: string
        senha:
          type: string

paths:
  /:
    get:
      summary: Rota principal de boas-vindas
      responses:
        '200':
          description: Informações da API
          content:
            application/json:
              schema:
                type: object
                properties:
                  api: { type: string }
                  version: { type: string }
                  Author: { type: string }

  /login:
    post:
      summary: Realiza login para obter o token de acesso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  token: { type: string }
        '401':
          description: Usuário ou senha inválidos

  /charadas:
    get:
      summary: Retorna todas as charadas
      responses:
        '200':
          description: Uma lista de charadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Charada'
    post:
      summary: Cria uma nova charada
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pergunta
                - resposta
              properties:
                pergunta: { type: string }
                resposta: { type: string }
      responses:
        '201':
          description: Charada criada com sucesso
        '400':
          description: Dados inválidos ou erro no envio

  /charadas/aleatoria:
    get:
      summary: Retorna uma charada aleatória
      responses:
        '200':
          description: Uma única charada aleatória
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charada'

  /charadas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID numérico da charada
        schema:
          type: integer
    get:
      summary: Busca uma charada pelo ID
      responses:
        '200':
          description: Dados da charada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charada'
        '404':
          description: Charada não encontrada
    
    put:
      summary: Atualização total de uma charada
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pergunta
                - resposta
              properties:
                pergunta: { type: string }
                resposta: { type: string }
      responses:
        '200':
          description: Alterada com sucesso
        '404':
          description: Não encontrada
        '400':
          description: Dados incompletos

    patch:
      summary: Atualização parcial de uma charada
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pergunta: { type: string }
                resposta: { type: string }
      responses:
        '200':
          description: Alterada com sucesso
        '404':
          description: Não encontrada

    delete:
      summary: Exclui uma charada
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Excluída com sucesso
        '404':
          description: Charada não encontrada